Create text field to text area
in this code we work the library swing and awt event for java we will make a graphical interface with a button to add text a field “field” to add what we write in this type of fields can not be made line break, we will make a variable of type text to host what we write in the first field and pass it to text area or textarea every time we press the button we will force the program to make a line break with \n.  we will use the content retrieval with getText and put the content elsewhere with setText. with getSource we tell the program to retrieve the event caused in the button1 when this happens we will tell it with += to accumulate the text that the user enters and so the conditional if will do its job we also put clean field at the end of the conditional if with (‘’);