Create JComboBox
This code was created for the menu combo box, which allows you to select multiple options. We created elements for the combo box, as shown in the code. In the public void itemStateChanged row, we can see the creation of the variable e to call the event when the options are pressed. The if condition returns everything inside the combo1 object. If it is correct, then later in the code: String select = combo1.getSelectedItem().toString(); 
We have to declare the variable of type String, its name is select, later we convert combo1 to type string within this, but first, we return here to getSelectItem(), this row is used to obtain the text or character of combo1 and with .toString() we can convert to string text. The next line is the title that uses the text within the select variable.